+++ apps/web/src/pages/get-started/consts/types.ts
@@ -1,0 +1,17 @@
+import { ProductUseCasesEnum } from '@novu/shared';
+
+export type OnboardingUseCases = {
+  [key in ProductUseCasesEnum]: OnboardingUseCase;
+};
+
+export interface IOnboardingStep {
+  title: string;
+  Description: React.ComponentType<any>;
+}
+
+export type OnboardingUseCase = {
+  title: string;
+  description: string;
+  steps: IOnboardingStep[];
+  Demo: React.ComponentType<any>;
+};
