+++ apps/web/src/pages/get-started/layout/GetStartedTab.tsx
@@ -1,16 +1,17 @@
 import React, { useEffect } from 'react';
 import { useNavigate, useParams } from 'react-router-dom';
 import { Grid } from '@mantine/core';
+import styled from '@emotion/styled';
 
 import { ProductUseCasesEnum } from '@novu/shared';
 import { colors, Text } from '@novu/design-system';
 
-import { onboardingUseCases } from '../consts';
 import { OnboardingParams } from '../types';
 import { ROUTES } from '../../../constants/routes.enum';
 import { OnboardingUseCasesTabsEnum } from '../../../constants/onboarding-tabs';
 import Card from '../../../components/layout/components/Card';
 import { Timeline } from '../components/timeline/Timeline';
+import { UseCasesConst } from '../consts/UseCases.const';
 
 interface IGetStartedTabProps {
   usecase?: ProductUseCasesEnum;
@@ -36,13 +37,13 @@ export function GetStartedTab(props: IGetStartedTabProps) {
     return null;
   }
 
-  const { steps, Demo, title, description } = onboardingUseCases[usecase];
+  const { steps, Demo, title, description } = UseCasesConst[usecase];
 
   return (
     <Grid align="stretch">
-      <Grid.Col span={4} mt={11.5}>
+      <Grid.Col span={4} mt={12}>
         <Card title={title} space={8} mb={24}>
-          <Text style={{ color: colors.B60 }}>{description}</Text>
+          <Description style={{ color: colors.B60 }}>{description}</Description>
         </Card>
         <Timeline steps={steps} />
       </Grid.Col>
@@ -53,3 +54,7 @@ export function GetStartedTab(props: IGetStartedTabProps) {
     </Grid>
   );
 }
+
+const Description = styled(Text)`
+  color: ${colors.B60};
+`;
